pyproject.toml
[project]
name = "rpaclaw"
version = "0.1.0"
description = "RPAclaw — Nanobot + RPA management platform with WebUI"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "RPAclaw contributors"}]
keywords = ["rpa", "automation", "ai", "agent", "webui"]

dependencies = [
    "nanobot-ai>=0.1.4",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "typer>=0.20.0,<1.0.0",
    "pydantic>=2.12.0,<3.0.0",
    "loguru>=0.7.3,<1.0.0",
    "robocorp-browser>=3.0.0",
    "robocorp-excel>=0.5.0",
    "rpaframework>=29.0.0",
    "rpaframework-pdf>=7.0.0",
]

[project.scripts]
rpaclaw = "rpaclaw.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rpaclaw"]
